+++ application/ticket-app-api/build.gradle
@@ -28,9 +28,8 @@ dependencies {
 	implementation project(":modules:jwt-provider")
 	implementation project(":modules:aws-s3")
 	implementation project(":modules:slack")
-	implementation project(":modules:token-redis")
 	implementation project(":modules:qrcode-provider")
-	implementation project(":modules:lock-redis")
+	implementation project(":modules:redis")
 
 	implementation project(":domain:form-domain")
 }
