+++ application/ticket-app-api/src/test/java/com/uket/app/ticket/api/TicketAppApiApplicationTests.java
@@ -1,12 +1,12 @@
 package com.uket.app.ticket.api;
 
-import com.uket.modules.redis.lock.config.RedissonConfig;
+import com.uket.modules.redis.config.RedisConfig;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 
 @SpringBootTest
-@Import({RedissonConfig.class})
+@Import({RedisConfig.class})
 class TicketAppApiApplicationTests {
 
     @Test
