--- modules/lock-redis/build.gradle
@@ -1,21 +1,0 @@
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-aop' // aop 추가
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-    implementation 'org.redisson:redisson-spring-boot-starter:3.30.0'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
-tasks.named('bootJar') {
-    enabled = false
-}
-
-tasks.named('jar') {
-    enabled = true
-}
