--- modules/lock-redis/src/main/java/com/uket/modules/redis/lock/aop/AopForTransaction.java
@@ -1,18 +1,0 @@
-package com.uket.modules.redis.lock.aop;
-
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * AOP에서 트랜잭션 분리를 위한 클래스
- */
-@Component
-public class AopForTransaction {
-
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public Object proceed(final ProceedingJoinPoint joinPoint) throws Throwable {
-        return joinPoint.proceed();
-    }
-}
