--- modules/lock-redis/src/main/java/com/uket/modules/redis/lock/properties/RedisLockProperties.java
@@ -1,11 +1,0 @@
-package com.uket.modules.redis.lock.properties;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "spring.data.redis.lock")
-public record RedisLockProperties(
-    String host,
-    int port
-) {
-
-}
