--- modules/lock-redis/src/main/java/com/uket/modules/redis/lock/util/CustomSpringELParser.java
@@ -1,25 +1,0 @@
-package com.uket.modules.redis.lock.util;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-import org.springframework.expression.ExpressionParser;
-import org.springframework.expression.spel.standard.SpelExpressionParser;
-import org.springframework.expression.spel.support.StandardEvaluationContext;
-
-/**
- * Spring Expression Language Parser
- */
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class CustomSpringELParser {
-
-    public static Object getDynamicValue(String[] parameterNames, Object[] args, String key) {
-        ExpressionParser parser = new SpelExpressionParser();
-        StandardEvaluationContext context = new StandardEvaluationContext();
-
-        for (int i = 0; i < parameterNames.length; i++) {
-            context.setVariable(parameterNames[i], args[i]);
-        }
-
-        return parser.parseExpression(key).getValue(context, Object.class);
-    }
-}
