+++ modules/redis/src/main/java/com/uket/modules/redis/aop/DistributedLockAop.java
@@ -1,0 +1,64 @@
+package com.uket.modules.redis.aop;
+
+import com.uket.modules.redis.util.CustomSpringELParser;
+import java.lang.reflect.Method;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.Around;
+import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.reflect.MethodSignature;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
+import org.springframework.stereotype.Component;
+
+/**
+ * @DistributedLock 선언 시 수행되는 Aop class
+ */
+@Aspect
+@Component
+@RequiredArgsConstructor
+@Slf4j
+public class DistributedLockAop {
+
+    private static final String REDISSON_LOCK_PREFIX = "LOCK:";
+
+    private final RedissonClient redissonClient;
+    private final AopForTransaction aopForTransaction;
+
+    @Around("@annotation(com.uket.modules.redis.aop.DistributedLock)")
+    public Object lock(final ProceedingJoinPoint joinPoint) throws Throwable {
+        MethodSignature signature = (MethodSignature) joinPoint.getSignature();
+        Method method = signature.getMethod();
+        DistributedLock distributedLock = method.getAnnotation(DistributedLock.class);
+
+        String key = REDISSON_LOCK_PREFIX + CustomSpringELParser.getDynamicValue(signature.getParameterNames(),
+                joinPoint.getArgs(), distributedLock.key());
+        RLock lock = redissonClient.getLock(key);
+
+        try {
+            boolean available = lock.tryLock(distributedLock.waitTime(), distributedLock.leaseTime(),
+                    distributedLock.timeUnit());
+
+            if (!available) {
+                return false;
+            }
+            return aopForTransaction.proceed(joinPoint);
+        } catch (InterruptedException e) {
+            throw new InterruptedException();
+        } finally {
+            try {
+                lock.unlock();   // (4)
+            } catch (IllegalMonitorStateException e) {
+                log.info("Redisson Lock Already UnLock {} {}",
+                        keyValue("serviceName", method.getName()),
+                        keyValue("key", key)
+                );
+            }
+        }
+    }
+
+    private Object keyValue(String key, String value) {
+        return String.join(":", key, value);
+    }
+}
