+++ modules/redis/src/main/java/com/uket/modules/redis/config/RedisConfig.java
@@ -1,0 +1,66 @@
+package com.uket.modules.redis.config;
+
+import com.uket.modules.redis.properties.RedisProperties;
+import lombok.RequiredArgsConstructor;
+import org.redisson.Redisson;
+import org.redisson.api.RedissonClient;
+import org.redisson.config.Config;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
+
+@Configuration
+@EnableRedisRepositories
+@RequiredArgsConstructor
+public class RedisConfig {
+    private final RedisProperties redisProperties;
+    private static final String REDISSON_HOST_PREFIX = "redis://";
+
+    //Redis Auth
+    public RedisProperties getRedisProperties() {
+        return redisProperties;
+    }
+    @Bean
+    public RedisConnectionFactory redisConnectionFactory() {
+        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration(redisProperties.host(), redisProperties.port());
+        return new LettuceConnectionFactory(config);
+    }
+
+    @Bean
+    public RedisTemplate<String, Object> redisTemplate() {
+        RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
+        redisTemplate.setConnectionFactory(redisConnectionFactory());
+
+        redisTemplate.setKeySerializer(new StringRedisSerializer());
+        redisTemplate.setValueSerializer(new StringRedisSerializer());
+
+        redisTemplate.setHashKeySerializer(new StringRedisSerializer());
+        redisTemplate.setHashValueSerializer(new StringRedisSerializer());
+
+        redisTemplate.setDefaultSerializer(new StringRedisSerializer());
+
+        return redisTemplate;
+    }
+
+    @Bean
+    public ValueOperations<String,String> valueOperations(RedisTemplate<String, String> redisTemplate){
+        return redisTemplate.opsForValue();
+    }
+
+    //Redis Lock
+    @Bean
+    public RedissonClient redissonClient() {
+
+        Config config = new Config();
+        String redisAddress = REDISSON_HOST_PREFIX + redisProperties.host() + ":" + redisProperties.port();
+        config.useSingleServer().setAddress(redisAddress);
+
+        return Redisson.create(config);
+    }
+}
