--- modules/token-redis/build.gradle
@@ -1,19 +1,0 @@
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-    //redis
-    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
-tasks.named('bootJar') {
-    enabled = false
-}
-
-tasks.named('jar') {
-    enabled = true
-}
