--- modules/token-redis/src/main/java/com/uket/modules/redis/constants/RedisValues.java
@@ -1,6 +1,0 @@
-package com.uket.modules.redis.constants;
-
-public class RedisValues {
-    public static final String REDIS_KEY_ACCESS_TOKEN = "accessToken";
-    public static final String REDIS_KEY_USER_ID = "userId";
-}
